% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutate.R
\name{weight_in_lbs}
\alias{weight_in_lbs}
\title{Convert weight from grams to pounds}
\usage{
weight_in_lbs(data = surveysnonas)
}
\arguments{
\item{data=surveysnonas}{Data that will be used in this function to get weight in pounds}
}
\value{
clean The data that has been cleaned/changed, in this case the weight column.
}
\description{
Multiply weight by 0.0022 in order to get pounds from grams
}
